+++ apps/web/next.config.js
@@ -1,7 +1,6 @@
 /** @type {import('next').NextConfig} */
 const nextConfig = {
   turbopack: {
-    root: new URL('../..', import.meta.url).pathname,
     rules: {
       '*.svg': {
         loaders: ['@svgr/webpack'],
