--- apps/web/src/apis/deleteWish.ts
@@ -1,5 +1,0 @@
-import { clientApi } from './client';
-
-export const deleteWish = async (wishId: number) => {
-  await clientApi.delete(`/api/v1/wishlists/${wishId}`);
-};
