+++ apps/web/src/app/wishlist/_apis/deleteWish.ts
@@ -1,0 +1,6 @@
+import { clientApi } from '@/apis/client';
+import { ENDPOINTS } from '@/consts/api';
+
+export const deleteWish = async (wishId: number) => {
+  await clientApi.delete(ENDPOINTS.WISHLIST(wishId));
+};
