+++ apps/web/src/app/wishlist/_components/WishTab.tsx
@@ -1,15 +1,34 @@
+'use client';
+
+import { useRouter, useSearchParams } from 'next/navigation';
+
 import { cn } from '@/utils/cn';
 
 import type { WishTabT } from '../_types/wishTypes';
 
 const TABS: WishTabT[] = ['저장한 위시템', '토너먼트 기록'];
 
-type WishTabProps = {
-  activeTab: WishTabT;
-  onTabChange: (tab: WishTabT) => void;
+const TAB_PARAM: Record<WishTabT, string> = {
+  '저장한 위시템': 'wish',
+  '토너먼트 기록': 'tournament',
 };
 
-function WishTab({ activeTab, onTabChange }: WishTabProps) {
+const PARAM_TAB: Record<string, WishTabT> = {
+  wish: '저장한 위시템',
+  tournament: '토너먼트 기록',
+};
+
+function WishTab() {
+  const router = useRouter();
+  const searchParams = useSearchParams();
+  const activeTab: WishTabT = PARAM_TAB[searchParams.get('tab') ?? ''] ?? '저장한 위시템';
+
+  const handleTabChange = (tab: WishTabT) => {
+    const params = new URLSearchParams(searchParams.toString());
+    params.set('tab', TAB_PARAM[tab]);
+    router.replace(`?${params.toString()}`);
+  };
+
   return (
     <div className="w-full px-5">
       <div className="flex h-[48px] items-center rounded-[12px] bg-[#E7E8EA] p-1">
@@ -17,7 +36,7 @@ function WishTab({ activeTab, onTabChange }: WishTabProps) {
           <button
             key={tab}
             type="button"
-            onClick={() => onTabChange(tab)}
+            onClick={() => handleTabChange(tab)}
             className={cn(
               'flex h-[40px] flex-1 items-center justify-center gap-2 text-[16.203px] leading-normal font-semibold transition-colors',
               activeTab === tab
