+++ apps/web/src/app/wishlist/_hooks/useDeleteWish.ts
@@ -1,6 +1,7 @@
 import { useMutation, useQueryClient } from '@tanstack/react-query';
+import { toast } from 'sonner';
 
-import { deleteWish } from '@/apis/deleteWish';
+import { deleteWish } from '../_apis/deleteWish';
 
 export const useDeleteWish = () => {
   const queryClient = useQueryClient();
@@ -9,6 +10,9 @@ export const useDeleteWish = () => {
     mutationFn: (wishId: number) => deleteWish(wishId),
     onSuccess: () => {
       queryClient.invalidateQueries({ queryKey: ['wishlists'] });
+    },
+    onError: () => {
+      toast.error('위시 삭제에 실패했어요');
     },
   });
 };
