+++ apps/web/src/app/wishlist/_hooks/useGetWishlist.ts
@@ -1,0 +1,10 @@
+import { useSuspenseQuery } from '@tanstack/react-query';
+
+import { getWishlist } from '../_apis/getWishlist';
+
+export const useGetWishlist = () => {
+  return useSuspenseQuery({
+    queryKey: ['wishlists'],
+    queryFn: getWishlist,
+  });
+};
