+++ apps/web/src/consts/api.ts
@@ -1,0 +1,4 @@
+export const ENDPOINTS = {
+  WISHLISTS: '/api/v1/wishlists',
+  WISHLIST: (id: number) => `/api/v1/wishlists/${id}`,
+} as const;
